FINISHED UI THAT GOT DESTROYED YESTERDAY
upgrade screen almost done and improved spawner

+++ Assets/Scripts/Health and Damage/Health.cs
@@ -26,7 +26,7 @@ public class Health : MonoBehaviour, IDamageable {
             OnDeath.Invoke();
         }
     }
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         HP += amount;
         if (HP > maxHP)
