+++ Assets/Scripts/Health and Damage/IDamageable.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamageable {
     void TakeDamage(float amount);
-    void Heal(int amount);
+    void Heal(float amount);
     void SetHP(float HP);
 }
